Added ability to delete or archive a cluster.
This only effects the visibility of the cluster.  It removes it from the main
cluster view.  It does not currently remove snapshots or any of the django
objects.  Perhaps there should be an option to remove snapshots.  There could be
recoverable and unrecoverable clusters where the presence of the snapshot
indicates whether it can be recovered.  I think the djano models should stick
around for historical reasons.